Extract `Glob` into a separate file and add doc
